Fixed walk command movement to be normalized properly.
Added the idea of "command properties" which enters the standardizing of them.  Thinking of a system
something like targetType, range, ect.

But for now I added "interuptable" to each command as a property.  Which allows you to interupt the current
command if another one is issued.

Mimics how it works in dota pretty well now.