After messing around with some circle shapes and lines I figured out that we need to have an "Object3D" parent that works as a "group" for our line.
This also means that we should just wrap everything in a parent Object3D, because
the functionally now exists to augment the model with things like the new selection circle ring.

There is an extra "selection circle" added to the scene group upon initalization with
the "visible = false" flag set.  When the entity is selected, the method "selectUnit(boolean selType)"
sets the circle visible and flags the entity as visible.  The EntityComponent also makes sure
that the previously selected unit is now not selected.

So now when you select units with the left mouse click the circle shows up,
and when they move they move their outer group so everything inside it
moves together!

Also probably going to rename some of the "Entity" class field names to make it
more obvious about the "parent object" and the main object meshes and such.

As it stands now you can use getMesh() to get the parent group, and 
getObjectMesh() for the main "object" (in our case its the square).  This
is used in the hit collision intersections, for whatever reason
the object3d mesh seems to not work but the objectmesh itself worked
great.  Could possibly not need this and might be using the parent
Object3D mesh wrong or something.