Changed overworld PerspectiveCamera arguments a bit.
Apparently "45 degrees" is what is considered "Natural" perspective so we'll use that.

Changed the camera position to (0,0,1000) for now.  Probably sticking with 1000 but could go back to 500.

Problems in this patch: the turning doesn't line up right for paths on diagonals.  The movement along an axis seems to do fine.

Maybe re-write the MoveCommand vector math to be simplier, as well as reform the angle fixing a little.