Major Restructuring ===================
This commit changes a few core classes and structure.

The EntityComponent still holds multiple instances of Entities.  The selection of Entities
will happen from an Event sent to the component from the SelectionComponent.

New Addition - entity/Entity.js
Location: G.entity.Entity

This is the new base "Entity" class.
The base Entity will hold two major objects inside:
	- The Model
	- The SceneObject

New Addition - sceneobject/SceneObject
Location: G.scene.SceneObject

This new class is the base object which is what is added directly to the openGL Scene.
It represents a Drawable scene graph object and allows for individual webGL libraries
to be abstracted from.

Defines core methods for Position, Rotation, and Scale for the objects represention
in the world scene.  Any class that extends SceneObject is responsible for
defining how it handles building and manipulating the scene graph by using the Prototype.js
method Object.extend() and overriding each blank SceneObject method with its own.

New Addition - sceneobject/THREESceneObject.js
Location: G.scene.THREESceneObject

A new class to represent the THREE.js library as a SceneObject.

The object is an extension of the THREE.Object3D class as well as
filling all the core methods of the SceneObject.

For this specific library we use the Object3D as our parent mesh
for the local group.  We are able to use the native methods
provided by the THREE.js library to add, remove, update,
and iterate through any added THREE.js scene graph objects.

The component "EntityComponent" for this game will produce a THREESceneObject
and a Model to represent an Entity in the world.

New Addition - controller/TestingController.js
Location: G.controller.TestingController

File used for debugging the engine.  When the G.testing flag is set to TRUE the
game.js file will default to loading this controller.  The controller
will spawn the TestingComponent as its only component.

Other components and controllers can also be added the usual way for testing.
This file and its component are meant as scratch for test work.

New Addition - component/TestingComponent.js
Location: G.component.TestingComponent

File used for debugging the engine.  When the G.testing flag is set to TRUE the
game.js file will default to loading this component as a part of the TestingController.