More core restructuring.  Got rid of some extra abstraction layers and simplified Entity and it's library specific extensions. Added base Model class and first actual model for Hero.
Entity will contain a Model, and a SceneObject which is defined in the library specific way. So in our three.js project
we have the file in entity/threejs/HeroEntity.js which contains THREE.js specific webGL definitions.

The EntityComponent will handle all the Entity instances in a given scene.  The Entity.SceneObject is added
to the opengl scene.  The Entity.Model is used for updates and sends event signals to the SceneObject,
and itself.

For now G.testing mode is enabled and the testcomponent is making an example entity in the scene.