Model default attribute fixes again
The base model class has defaults that are loaded always.

But if you pass options to the extended Model (in our case HeroModel)
those will also use _.defaults() and then pass that into our base Model
to use _.defaults().

This allows you to optionally overwrite a default value in a Model.