EntityFactory, Settings, SceneGraphs ========================================================= The entity class is now a singular class which accepts a Type, a Model, and a SceneGraph.
Settings
=========================================================

The Settings object represents each entity type as a model and a scene graph.
Planned expansion for animations described later in message.

SceneGraph
===========================================================
SceneGraph is the terminology used to defined what is represented by three.js as an "Object3D",
the base class for everything else made in the library.

Because the scene graph is the portion of the code where we define library specific 
ways of representing them, we create a version of each scene graph as a new class
with the "buildScene" method that accepts a model.

The model should be able to define characteristics of the scenegraph, such as color,
textures, and other attributes such as size.

In the future we may need to consider the animation of these scene graph objects
and arrange it so that the settings of an Entity points to a group of
scene objects represented by the animation.  For example a walk animation
will consist of a series of scene graphs.  The settings can point to an "animation list"
of these individual scene graph parts.

Warning System
====================================================

Also changed the warning system a bit.  The new flag "showWarnings" will enable
console logging of the warnings.

When in "testing" mode, the warnings are set to be shown by default.