Created base SceneGraph, moved events to SceneGraph class.
We might want a "SceneGraph" class that isn't threejs specific.  We could have a
base SceneGraph and then a THREESceneGraph which just fills in the methods like
"setX" and "setRotationX" and such.

As of right now it just creates thee base SceneGraph class within the threejs folder, which is
a working solution for now.

Also moved events to this class, and switched it so the SceneGraph extensions define a
method called "setup", and the base class gets "buildScene" called.  This way we can
build the scenegraph with the model, and then set the default scene options
such as starting (x,y,z) and rotations and such.

Instead of doing that we could just fire the events for each starting attribute as well.