Created new OverworldComponent.  Added listeners in EntityComponent ====================================================
The new component is going to be the main component going forward. Thinking it will hold
the team ID's and neutral information and such.

Next part is to add the loader class so we can get some dynamic entity loading going on.
Going to need basic things like the ground for testing / selection development.