Changed the loader to an "ObjectLoader" which is what is 'supposed' to load the exports from the three editor.
So now that loading seems to work just fine we need to decide how the loading is
going to function.

The case for making it a controller is such that we can pass in a List or even a 
JSON file which contains everything that needs to be loaded by the loader.

So you can still have a game where you only load a partial of the world and
go through doors and have secondary load screens.

It simplifies the interaction of waiting for the load as well, just point the router
to the Loader controlller, and when the load is done the loading controller
tells the router to get what comes next.