Added a loader instance to Router
The router now has a single instace of the loader and will go through a life cycle with each controller.

The router attaches a callback method called "loadComplete" to the loader which is called upon
load()'s completion.

Next step is defining the resources that each controller has and loading them as a list of assets.