Introduced resource array into controllers and components, Implemented LoaderCache ========================================================
Resources - Controllers and components, Controller.getResources(), Router Changes
========================================================
The controllers are able to specific resoruces they want loaded (can just take that out if 
they never do), but they also load every components listed resoruce as well.

The code uses the Array.prototype.concat() function to merge the two Arrays.  This
function also has the feature where it doesn't add the same string entry twice
into the result Array.  This serves as a way for the controller and it's components
to not reload the same url string more than once.

The router now handles loading more than one url resource.  This is handled
by asking the controler after the init() phase for the return of
Controller.getResources().  This list and the length of the list are used
to determine when the Controller is completely finished loading.

LoaderCache
===================================================

The LoaderCache class is an interface for a HashTable of entries that have
already been loaded by the given Loader.

The cache will add an entry for a URL by giving a "null" value for the entry
in the HashTable.

The callback function for the loader must be able to identify the URL of the
result from the itemEnd() method.  This is done by passing the variable
into an anon. function with the url coded in.