Major Changes ==========================================
This patch contains some major changes in that it adds resource loading capability and introduces
the global dispatcher (gloal event dispatcher).

Controller and Component Changes
======================================================

Controllers and Components will now both listen for resource related events.  Currently
the only event is "LOADER.Finish" (pending naming scheme). This event is fired by the
global dispatch and all components and controllers will get the event.

When a controller is requested in the Router, the Router will use it's new Loader instance
to call "loadResources" on the controller.  This function checks the "resources" Object of
the Controller and each of it's components and compiles a unique Array of URL entries
to load.

The Loader.load function accepts a callback, which simply fires an event with the global dispatch named
"LOADING.Finish" as we saw from before.

Currently both Controllers and Components have a "resourceCount" property generated at their initalization.
This number is used to check if there are any still items left to be loaded.


**********************************************************************************
TODO: 

Eventually the plan is to remove the global listener when everything has been loaded.
But there were some weird scope issues with removing the listener and such... so i might need
just some kind of non-anonymous function for the LOADER.Finish callback so that it can be removed. 

The cache system does not currently use a value that may be stored in it.  It has the functionality
just this patch didn't hook up the cache yet.
**********************************************************************************