Reverted back to using THREE.LoadingManager, added setLoaderCache to inject cache into controllers.
Some of the router functionality for a controller that isn't the "current" focus is there with the swap argument.
Haven't had to handle a controller made like this so for now it'll stay like this. Shouldn't be hard to extend the router
more.

Next up is hooking up the cache fully, handling controllers with no resoruces to load, handling situations where you get
nothing but cache hits, and such.