Introduced ResourceBank
The name is up for change ofc, but i figured we wanted to use "resources" as the reserve for what to preload.  Could just change that to like "preloadResources" and this to G.resources.Resources() or something.

Changed the ThreeLoader to include an item finish callback function for the manager.  Now the router handles when something finishes loading. 

TODO: There are some arguments in the itemprogress that need to be checked as well once error handling is in there.