Todo upload
Quick scribble:

Local resource names shouldn't be a thing, doesn't make sense for the buildScene calls.  It should just be a 
G.resourceBank and the controllers / components will just list an array of names that they need for loading.

So in a buildScene you'd just say this.container.add(G.resourceBank.get("GroundObject")); and you'd
push the object into the scene.