Changed Controller.loadResource to fit ResourceBank data scheme
Now the loadResource function computes the _.pairs() ( [ LocalName, URL ] ) of the resource object
for the controller and it's components, and takes a unique result of that Array utilizing the
optional iterator function in _.unique() to ensure that no two resources have the same local name.

Resources can however declare the same url but with a different localized name.  This is okay
because a collision of the local name would stop the ResourceBank.get() calls from working.
The ResourceBank should not need to provide a getter based on the URL in theory.
(i think theres still a function in there for it ? maybe?)  The cache should still have that
functionality though of course.