Major clean-up.  TestController now shows off an Entity on click.
The Entity now has a SceneGraph which can utilize a resource in the ResourceBank of which
is an object defined by the three.js editor.

Currently when you click on the screen a plane is added, which is defined as the Ground entity
using the Ground SceneGraph (just a plane).